SWM-35 refactor: DiarySummary 분리
