+++ app/src/main/java/com/betterlife/antifragile/data/model/diary/DiarySummary.kt
@@ -1,0 +1,7 @@
+package com.betterlife.antifragile.data.model.diary
+
+data class DiarySummary(
+    val id: Int,
+    val date: String,
+    val emotionIconUrl: String?
+)
